Migrated NoContentSerializer tests.
